Move tool_repair.py out of long_running/
``databricks_langchain.checkpoint`` is Python 3.10+ and re-exported
eagerly from ``databricks_langchain/__init__.py``. Placing
``DEFAULT_SYNTHETIC_INTERRUPTED_OUTPUT`` inside ``long_running/repair.py``
meant any ``import databricks_langchain`` triggered
``long_running/__init__.py``, which:

1. ``from .server import ...`` — ``server.py`` raises RuntimeError on
   Python < 3.11, breaking langchain tests on 3.10.
2. ``from .db import ...`` — pulls ``psycopg``, breaking any langchain
   test that doesn't have ``[memory]`` extras installed.

Both failed in CI. Move the file to top-level
``databricks_ai_bridge.tool_repair`` so it can be imported without
crossing the ``long_running/`` package boundary. Three import sites
updated (server, session, checkpoint) plus the test. No API break —
``sanitize_tool_items`` and ``DEFAULT_SYNTHETIC_INTERRUPTED_OUTPUT``
were still pre-release.
